requirements.txt
python-telegram-bot==20.6
APScheduler==3.11.0
